ref: move client tests to "tests" dir
